# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: atx-perform API
  version: "1.0"
  description: |
    前后端分离契约。Demo 轨由 FileSceneRepository 读 data/1-*.json；
    Live 轨（PG / DeepAgent / 技能落盘）为模块占位，返回 NOT_IMPLEMENTED。
servers:
  - url: http://127.0.0.1:8000/api/v1
paths:
...
  /scenes/{scene_key}:
    get:
      operationId: getScene
      summary: 按幕加载数据集
      parameters:
        - in: path
          name: scene_key
          required: true
          schema: { type: string, enum: ["0", "1", "2", "3", "4", "5"] }
      responses:
        "200":
          description: 幕数据包
          content:
            application/json:
              schema: { $ref: "./schemas/scene-bundle.json" }
        "404":
          description: 未知幕
          content:
            application/json:
              schema: { $ref: "./schemas/error.json" }
  /scenes/bundle:
    get:
      operationId: getSceneBundle
      summary: 按数据集名加载（对齐 loadSceneBundle）
      parameters:
        - in: query
          name: names
          required: true
          schema: { type: string }
          description: 逗号分隔，如 locate,channelization
      responses:
        "200":
          description: 数据集对象
          content:
            application/json:
              schema:
                type: object
                required: [source, contract_version, generated_at, datasets]
                properties:
                  source: { type: string, enum: [demo, live] }
                  contract_version: { type: string }
                  generated_at: { type: string, format: date-time }
                  datasets:
                    type: object
                    additionalProperties: { type: object }
        "400":
          description: 非法数据集名
          content:
            application/json:
              schema: { $ref: "./schemas/error.json" }
...